+++ client/src/shared/components/UIElements/FormElements/components/FormikFieldArray.js
@@ -1,57 +1,95 @@
 import React from 'react';
-import {
-  Input,
-  FormControl,
-  FormLabel,
-  Button,
-  FormErrorMessage,
-} from '@chakra-ui/react';
-import { Field, Form, Formik } from 'formik';
+import { Formik, Form, FieldArray } from 'formik';
+import { HStack, VStack, Button, Heading } from '@chakra-ui/react';
 
-function FormikFieldArray() {
-  function validateName(value) {
-    let error;
-    if (!value) {
-      error = 'Name is required';
-    } else if (value.toLowerCase() !== 'naruto') {
-      error = "Jeez! You're not a fan 😱";
-    }
-    return error;
-  }
+import { validateRequire } from './FormikValidations';
 
-  return (
+import FormikInput from './FormikInput';
+
+const initialValues = {
+  env: [
+    {
+      var: '',
+      val: '',
+    },
+  ],
+};
+
+const FormikFieldArray = () => (
+  <VStack
+    p="4"
+    m="8"
+    boxShadow="md"
+    // borderColor="gray.200"
+    // borderWidth="2px"
+    borderRadius="xl"
+    // w="50%"
+    w="500px"
+    // maxW={{ base: '90vw', sm: '80vw', lg: '50vw', xl: '40vw' }}
+    alignItems="stretch"
+  >
+    <Heading align="center" size="lg">
+      Add ENV Variables
+    </Heading>
+
     <Formik
-      initialValues={{ name: 'Sasuke' }}
-      onSubmit={(values, actions) => {
-        setTimeout(() => {
-          alert(JSON.stringify(values, null, 2));
-          actions.setSubmitting(false);
-        }, 1000);
+      initialValues={initialValues}
+      onSubmit={async values => {
+        await new Promise(r => setTimeout(r, 500));
+        alert(JSON.stringify(values, null, 2));
       }}
     >
-      {props => (
+      {({ values }) => (
         <Form>
-          <Field name="name" validate={validateName}>
-            {({ field, form }) => (
-              <FormControl isInvalid={form.errors.name && form.touched.name}>
-                <FormLabel htmlFor="name">First name</FormLabel>
-                <Input {...field} id="name" placeholder="name" />
-                <FormErrorMessage>{form.errors.name}</FormErrorMessage>
-              </FormControl>
+          <FieldArray name="env">
+            {({ insert, remove, push }) => (
+              <VStack>
+                {values.env.length > 0 &&
+                  values.env.map((friend, index) => (
+                    <HStack key={index}>
+                      <FormikInput
+                        uniqueField={`env.${index}.var`}
+                        validation={validateRequire}
+                        // label="Name"
+                        placeholder="DOMAIN_NAME"
+                        type="text"
+                      />
+
+                      <FormikInput
+                        uniqueField={`env.${index}.val`}
+                        validation={validateRequire}
+                        // label="Email"
+                        placeholder="app.example.com"
+                        type="text"
+                      />
+                      <Button
+                        type="button"
+                        className="secondary"
+                        onClick={() => push({ var: '', val: '' })}
+                      >
+                        +
+                      </Button>
+
+                      {values.env.length !== 1 && (
+                        <Button
+                          type="button"
+                          className="secondary"
+                          onClick={() => remove(index)}
+                        >
+                          x
+                        </Button>
+                      )}
+                    </HStack>
+                  ))}
+              </VStack>
             )}
-          </Field>
-          <Button
-            mt={4}
-            colorScheme="teal"
-            isLoading={props.isSubmitting}
-            type="submit"
-          >
-            Submit
-          </Button>
+          </FieldArray>
+          <Button type="submit">Save</Button>
         </Form>
       )}
     </Formik>
-  );
-}
+  </VStack>
+);
 
 export default FormikFieldArray;
+// ReactDOM.render(<InviteFriends />, document.getElementById('root'));
